back_end/requirements.txt
django~=3.0.0
djangorestframework~=3.11